refactor[structure]: project structure modify
