+++ components/schedule/Schedule.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Schedule = () => {
+    return (
+        <div>
+            schedule component for example 
+        </div>
+    );
+};
+
+export default Schedule;
